Message filter - simple string doesn't require properties to be in header
