+++ src/Common/Helpers/LogBrokeredMessageInspector.cs
@@ -26,10 +26,8 @@ using System.Collections.Concurrent;
 using System.Globalization;
 using System.IO;
 using System.Linq;
-using System.ServiceModel.Channels;
 using System.Text;
 using System.Threading.Tasks;
-using System.Xml;
 using Microsoft.ServiceBus.Messaging;
 
 #endregion
@@ -47,14 +45,12 @@ namespace ServiceBusExplorer.Helpers
         private const string DateFormat = "<{0,2:00}:{1,2:00}:{2,2:00}> {3}";
         private const string MessageSuccessfullySent = "Message Sent. MessageId=[{0}] SessionId=[{1}] Label=[{2}] Size=[{3}]";
         private const string MessageSuccessfullyReceived = "Message Received. MessageId=[{0}] SessionId=[{1}] Label=[{2}] Size=[{3}]";
-        private const string UnableToReadMessageBody = "Unable to read the message body.";
         private const string NullValue = "NULL";
         private const string MessagePropertiesHeader = "Properties:";
         private const string MessagePayloadHeader = "Payload:";
         private const string MessageTextFormat = "{0}";
         private const string MessagePropertyFormat = " - Key=[{0}] Value=[{1}]";
         private const string LogFileNameFormat = "BrokeredMessage {0}.txt";
-        private const int MaxBufferSize = 262144; // 256 KB
         #endregion
 
         #region Private Instance Fields
@@ -116,134 +112,6 @@ namespace ServiceBusExplorer.Helpers
             return message;
         }
 
-        /// <summary>
-        /// Reads the content of the BrokeredMessage passed as argument.
-        /// </summary>
-        /// <param name="messageToRead">The BrokeredMessage to read.</param>
-        /// <returns>The content of the BrokeredMessage.</returns>
-        private static string GetMessageText(BrokeredMessage messageToRead)
-        {
-            string messageText = null;
-            Stream stream = null;
-            
-            if (messageToRead == null)
-            {
-                return null;
-            }
-            var inboundMessage = messageToRead.Clone();
-            try
-            {
-                stream = inboundMessage.GetBody<Stream>();
-                if (stream != null)
-                {
-                    var element = new BinaryMessageEncodingBindingElement
-                    {
-                        ReaderQuotas = new XmlDictionaryReaderQuotas
-                        {
-                            MaxArrayLength = int.MaxValue,
-                            MaxBytesPerRead = int.MaxValue,
-                            MaxDepth = int.MaxValue,
-                            MaxNameTableCharCount = int.MaxValue,
-                            MaxStringContentLength = int.MaxValue
-                        }
-                    };
-                    var encoderFactory = element.CreateMessageEncoderFactory();
-                    var encoder = encoderFactory.Encoder;
-                    var stringBuilder = new StringBuilder();
-                    var message = encoder.ReadMessage(stream, MaxBufferSize);
-                    using (var reader = message.GetReaderAtBodyContents())
-                    {
-                        // The XmlWriter is used just to indent the XML message
-                        var settings = new XmlWriterSettings { Indent = true };
-                        using (var xmlWriter = XmlWriter.Create(stringBuilder, settings))
-                        {
-                            xmlWriter.WriteNode(reader, true);
-                        }
-                    }
-                    messageText = stringBuilder.ToString();
-                }
-            }
-            catch (Exception)
-            {
-                inboundMessage = messageToRead.Clone();
-                try
-                {
-                    stream = inboundMessage.GetBody<Stream>();
-                    if (stream != null)
-                    {
-                        var element = new BinaryMessageEncodingBindingElement
-                        {
-                            ReaderQuotas = new XmlDictionaryReaderQuotas
-                            {
-                                MaxArrayLength = int.MaxValue,
-                                MaxBytesPerRead = int.MaxValue,
-                                MaxDepth = int.MaxValue,
-                                MaxNameTableCharCount = int.MaxValue,
-                                MaxStringContentLength = int.MaxValue
-                            }
-                        };
-                        var encoderFactory = element.CreateMessageEncoderFactory();
-                        var encoder = encoderFactory.Encoder;
-                        var message = encoder.ReadMessage(stream, MaxBufferSize);
-                        using (var reader = message.GetReaderAtBodyContents())
-                        {
-                            messageText = reader.ReadString();
-                        }
-                    }
-                }
-                catch (Exception)
-                {
-                    try
-                    {
-                        if (stream != null)
-                        {
-                            try
-                            {
-                                stream.Seek(0, SeekOrigin.Begin);
-                                var serializer = new CustomDataContractBinarySerializer(typeof(string));
-                                messageText = serializer.ReadObject(stream) as string;
-                            }
-                            catch (Exception)
-                            {
-                                try
-                                {
-                                    stream.Seek(0, SeekOrigin.Begin);
-                                    using (var reader = new StreamReader(stream))
-                                    {
-                                        messageText = reader.ReadToEnd();
-                                        if (messageText.ToCharArray().GroupBy(c => c).
-                                            Where(g => char.IsControl(g.Key) && g.Key != '\t' && g.Key != '\n' && g.Key != '\r').
-                                            Select(g => g.First()).Any())
-                                        {
-                                            stream.Seek(0, SeekOrigin.Begin);
-                                            using (var binaryReader = new BinaryReader(stream))
-                                            {
-                                                var bytes = binaryReader.ReadBytes((int)stream.Length);
-                                                messageText = BitConverter.ToString(bytes).Replace('-', ' ');
-                                            }
-                                        }
-                                    }
-                                }
-                                catch (Exception)
-                                {
-                                    messageText = UnableToReadMessageBody;
-                                }
-                            }
-                        }
-                        else
-                        {
-                            messageText = UnableToReadMessageBody;
-                        }
-                    }
-                    catch (Exception)
-                    {
-                        messageText = UnableToReadMessageBody;
-                    }
-                }
-            }
-            return messageText;
-        }
-
         private async static void WriteToLog()
         {
             try
@@ -273,7 +141,7 @@ namespace ServiceBusExplorer.Helpers
                                 string.IsNullOrWhiteSpace(message.Label) ? NullValue : message.Label,
                                 message.Size));
                             builder.AppendLine(MessagePayloadHeader);
-                            var messageText = GetMessageText(message);
+                            var messageText = message.GetMessageText();
                             builder.AppendLine(string.Format(MessageTextFormat,
                                 messageText.Contains('\n')
                                     ? messageText
