+++ src/ServiceBusExplorer/Controls/ListenerControl.cs
@@ -1182,6 +1182,10 @@ namespace ServiceBusExplorer.Controls
                 if (receiverBrokeredMessageInspector != null)
                 {
                     message = receiverBrokeredMessageInspector.AfterReceiveMessage(message);
+                    if (message == null)
+                    {
+                        return;
+                    }
                 }
                 if (logging)
                 {
